+++ src/routes.rs
@@ -12,7 +12,7 @@ pub fn get_all_posts(conn: DbConn) -> Json<Value> {
 
 
 #[get("/posts/<id>")]
-pub fn get_post(id: i32, conn: DbConn) -> Json<Value> {
+pub fn get_post(id: i64, conn: DbConn) -> Json<Value> {
     Json(json!({"status": 200, "posts": Post::get(id, &conn)}))
 }
 
@@ -24,7 +24,7 @@ pub fn create_post(post: Json<NewPost>, conn: DbConn) -> Json<Value> {
 
 
 #[patch("/posts/<id>", format = "application/json", data = "<post>")]
-pub fn update_post(id: i32, post: Json<NewPost>, conn: DbConn) -> Result<Json<Value>, Status> {
+pub fn update_post(id: i64, post: Json<NewPost>, conn: DbConn) -> Result<Json<Value>, Status> {
     let updated = Post::update(&conn, id, &post.into_inner());
     match updated {
         true => Ok(Json(json!({"status": 200, "updated": updated}))),
@@ -34,7 +34,7 @@ pub fn update_post(id: i32, post: Json<NewPost>, conn: DbConn) -> Result<Json<Va
 
 
 #[delete("/posts/<id>")]
-pub fn delete_post(id: i32, conn: DbConn) -> Json<Value> {
+pub fn delete_post(id: i64, conn: DbConn) -> Json<Value> {
     Json(json!({"status": 200, "deleted": Post::delete(&conn, id)}))
 }
